+++ backend-spring-boot/src/main/java/com/sword/trailersdb/exceptions/ActorNotFoundException.java
@@ -1,4 +1,4 @@
-package com.sword.trailersdb.exception;
+package com.sword.trailersdb.exceptions;
 
 public class ActorNotFoundException extends RuntimeException {
     public ActorNotFoundException(Long id) {
